test: extract bash classifier white-box suite
